requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coala-utils==0.5.1
fake-useragent==0.1.11
fastapi==0.31.0
h11==0.8.1
httptools==0.0.13
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
pycodestyle==2.5.0
pydantic==0.29
pylint==2.3.1
pymongo==3.8.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
starlette==0.12.0
twitterscraper==1.2.0
typed-ast==1.4.0
urllib3==1.25.3
uvicorn==0.8.3
uvloop==0.12.2
websockets==7.0
wrapt==1.11.2
